+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added check for user being its own manager in SF1500.
+
 ## [1.3.0] - 2023-02-03
 
 ### Added